Model_Integration/requirements.txt
Flask>=3.0.0
google-generativeai>=0.8.5
google-adk>=1.8.0
google-cloud-storage>=2.0.0
